main_max_prefix_sum: fix bug in data generation

+++ src/main_max_prefix_sum.cpp
@@ -28,7 +28,7 @@ int main(int argc, char **argv)
         std::vector<int> as(n, 0);
         FastRandom r(n);
         for (int i = 0; i < n; ++i) {
-            as[i] = (unsigned int) r.next(-values_range, values_range);
+            as[i] = r.next(-values_range, values_range);
         }
 
         int reference_max_sum;
